# simpler-api/openapi.yaml
openapi: 3.1.0
info:
  contact:
    email: philipp.schmurr@kit.edu
  description: "This is a Schema extraction API based on the OpenAPI 3.1 specification.\
    \  You can find out more about\nSwagger at [https://swagger.io](https://swagger.io). "
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: Schema API - OpenAPI 3.1
  version: 0.7.0
...
servers:
- url: http://localhost:7373
...
paths:
...
  /schemata:
    get:
      description: desc
      operationId: getAllSchemas
      parameters:
      - $ref: '#/components/parameters/preventOptimization'
      - $ref: '#/components/parameters/preventAutomaticOptimization'
      - $ref: '#/components/parameters/preventUserOptimization'
      - $ref: '#/components/parameters/generateInverseRelations'
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Schema'
                type: array
            application/xml:
              schema:
                items:
                  $ref: '#/components/schemas/Schema'
                type: array
                xml:
                  name: schemas
            application/x-yaml:
              schema:
                items:
                  $ref: '#/components/schemas/Schema'
                type: array
          description: successful operation
      summary: Get all available schemas from this API server
      tags:
      - schema
  /schemata/{schemaId}:
    get:
      description: desc
      operationId: getSchemaByID
      parameters:
      - $ref: '#/components/parameters/schemaId'
      - $ref: '#/components/parameters/preventOptimization'
      - $ref: '#/components/parameters/preventAutomaticOptimization'
      - $ref: '#/components/parameters/preventUserOptimization'
      - $ref: '#/components/parameters/generateInverseRelations'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schema'
            application/x-yaml:
              schema:
                $ref: '#/components/schemas/Schema'
            application/xml:
              schema:
                $ref: '#/components/schemas/Schema'
          description: successful operation
        "404":
          description: not found
      summary: Get schema by ID
      tags:
      - schema
...
  /schemata/{schemaId}/diagram:
    get:
      description: desc
      operationId: getSchemaDiagram
      parameters:
      - $ref: '#/components/parameters/schemaId'
      - description: Toggle the rendering state of attributes
        explode: true
        in: query
        name: showAttributes
        required: false
        schema:
          default: true
          type: boolean
        style: form
      - description: Specify a list of highlighted entities
        explode: true
        in: query
        name: selectedEntities
        required: false
        schema:
          default: []
          items:
            type: string
          type: array
        style: form
      - description: The distance to render when using highlighted entities
        explode: true
        in: query
        name: renderDistance
        required: false
        schema:
          default: 2
          type: integer
        style: form
      - $ref: '#/components/parameters/preventOptimization'
      - $ref: '#/components/parameters/preventAutomaticOptimization'
      - $ref: '#/components/parameters/preventUserOptimization'
      - $ref: '#/components/parameters/generateInverseRelations'
      responses:
        "200":
          content:
            image/svg+xml:
              schema:
                type: string
            text/plain:
              schema:
                type: string
          description: successful operation
        "404":
          description: schema not found
      summary: Get a diagram for the schema
      tags:
      - schema
  /schemata/{schemaId}/partitioned-entities:
    get:
      description: desc
      operationId: getPartitionedEntitiesBySchema
      parameters:
      - $ref: '#/components/parameters/schemaId'
      - $ref: '#/components/parameters/preventOptimization'
      - $ref: '#/components/parameters/preventAutomaticOptimization'
      - $ref: '#/components/parameters/preventUserOptimization'
      - $ref: '#/components/parameters/generateInverseRelations'
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Partition'
                type: array
            application/x-yaml:
              schema:
                items:
                  $ref: '#/components/schemas/Partition'
                type: array
            application/xml:
              schema:
                items:
                  $ref: '#/components/schemas/Partition'
                type: array
                xml:
                  name: Partitions
          description: successful operation
      tags:
      - partition
  /schemata/{schemaId}/entities:
    get:
      description: desc
      operationId: getEntitiesBySchema
      parameters:
      - $ref: '#/components/parameters/schemaId'
      - $ref: '#/components/parameters/preventOptimization'
      - $ref: '#/components/parameters/preventAutomaticOptimization'
      - $ref: '#/components/parameters/preventUserOptimization'
      - $ref: '#/components/parameters/generateInverseRelations'
      - $ref: '#/components/parameters/entityFilter'
      - $ref: '#/components/parameters/entityFilterDistance'
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Entity'
                type: array
            application/xml:
              schema:
                items:
                  $ref: '#/components/schemas/Entity'
                type: array
                xml:
                  name: entities
            application/x-yaml:
              schema:
                items:
                  $ref: '#/components/schemas/Entity'
                type: array
          description: successful operation
      summary: Get all entities of a schema
      tags:
      - entity
  /schemata/{schemaId}/entities/{entityId}:
    get:
      description: desc
      operationId: getEntityById
      parameters:
      - $ref: '#/components/parameters/schemaId'
      - $ref: '#/components/parameters/entityId'
      - $ref: '#/components/parameters/preventOptimization'
      - $ref: '#/components/parameters/preventAutomaticOptimization'
      - $ref: '#/components/parameters/preventUserOptimization'
      - $ref: '#/components/parameters/generateInverseRelations'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entity'
            application/xml:
              schema:
                $ref: '#/components/schemas/Entity'
            application/x-yaml:
              schema:
                $ref: '#/components/schemas/Entity'
          description: successful operation
      summary: Get a specific entity
      tags:
      - entity
components:
  parameters:
...
    preventOptimization:
      explode: true
      in: query
      name: preventOptimization
      required: false
      schema:
        default: false
        type: boolean
      style: form
    preventAutomaticOptimization:
      explode: true
      in: query
      name: preventAutomaticOptimization
      required: false
      schema:
        default: false
        type: boolean
      style: form
    preventUserOptimization:
      explode: true
      in: query
      name: preventUserOptimization
      required: false
      schema:
        default: false
        type: boolean
      style: form
    generateInverseRelations:
      description: Toggles whether to generate inverse relations for each existing
        relation that has no schema based inverse
      explode: true
      in: query
      name: generateInverseRelations
      required: false
      schema:
        default: false
        type: boolean
      style: form
    entityFilter:
      description: Only list entities of the schema that are in range of the specified
        entities
      explode: true
      in: query
      name: entityFilter
      required: false
      schema:
        default: []
        items:
          type: string
        type: array
      style: form
    entityFilterDistance:
      description: Specifies the search distance if entityFilter is used
      explode: true
      in: query
      name: entityFilterDistance
      required: false
      schema:
        default: 1
        type: integer
      style: form
    entityId:
      description: ID of entity to access
      explode: false
      in: path
      name: entityId
      required: true
      schema:
        type: string
      style: simple
    schemaId:
      description: ID of schema to access
      explode: false
      in: path
      name: schemaId
      required: true
      schema:
        type: string
      style: simple
...
  schemas:
...
    EntityLink:
      format: url
      title: EntityLink
      type: string
...
    Partition:
      example:
        entities:
        - isObjectInRelation:
          - objectCardinality:
              cardinality: cardinality
            subjectCardinality:
              cardinality: cardinality
            hasSubjectEntity: ""
            relationName:
            - relationName
            - relationName
            hasObjectEntity: ""
            inverseRelation: inverseRelation
            hasAttribute:
            - attributeName:
              - attributeName
              - attributeName
              hasAttributeModifier:
              - attributeModifier: attributeModifier
              - attributeModifier: attributeModifier
            - attributeName:
              - attributeName
              - attributeName
              hasAttributeModifier:
              - attributeModifier: attributeModifier
              - attributeModifier: attributeModifier
            hasRelationModifier:
            - relationModifier: relationModifier
            - relationModifier: relationModifier
          - objectCardinality:
              cardinality: cardinality
            subjectCardinality:
              cardinality: cardinality
            hasSubjectEntity: ""
            relationName:
            - relationName
            - relationName
            hasObjectEntity: ""
            inverseRelation: inverseRelation
            hasAttribute:
            - attributeName:
              - attributeName
              - attributeName
              hasAttributeModifier:
              - attributeModifier: attributeModifier
              - attributeModifier: attributeModifier
            - attributeName:
              - attributeName
              - attributeName
              hasAttributeModifier:
              - attributeModifier: attributeModifier
              - attributeModifier: attributeModifier
            hasRelationModifier:
            - relationModifier: relationModifier
            - relationModifier: relationModifier
          hasEntityModifier:
          - entityModifier: entityModifier
          - entityModifier: entityModifier
          entityName:
          - entityName
          - entityName
          entityUrl: entityUrl
          hasAttribute:
          - attributeName:
            - attributeName
            - attributeName
            hasAttributeModifier:
            - attributeModifier: attributeModifier
            - attributeModifier: attributeModifier
          - attributeName:
            - attributeName
            - attributeName
            hasAttributeModifier:
            - attributeModifier: attributeModifier
            - attributeModifier: attributeModifier
          isSubjectInRelation:
          - objectCardinality:
              cardinality: cardinality
            subjectCardinality:
              cardinality: cardinality
            hasSubjectEntity: ""
            relationName:
            - relationName
            - relationName
            hasObjectEntity: ""
            inverseRelation: inverseRelation
            hasAttribute:
            - attributeName:
              - attributeName
              - attributeName
              hasAttributeModifier:
              - attributeModifier: attributeModifier
              - attributeModifier: attributeModifier
            - attributeName:
              - attributeName
              - attributeName
              hasAttributeModifier:
              - attributeModifier: attributeModifier
              - attributeModifier: attributeModifier
            hasRelationModifier:
            - relationModifier: relationModifier
            - relationModifier: relationModifier
          - objectCardinality:
              cardinality: cardinality
            subjectCardinality:
              cardinality: cardinality
            hasSubjectEntity: ""
            relationName:
            - relationName
            - relationName
            hasObjectEntity: ""
            inverseRelation: inverseRelation
            hasAttribute:
            - attributeName:
              - attributeName
              - attributeName
              hasAttributeModifier:
              - attributeModifier: attributeModifier
              - attributeModifier: attributeModifier
            - attributeName:
              - attributeName
              - attributeName
              hasAttributeModifier:
              - attributeModifier: attributeModifier
              - attributeModifier: attributeModifier
            hasRelationModifier:
            - relationModifier: relationModifier
            - relationModifier: relationModifier
        - isObjectInRelation:
          - objectCardinality:
              cardinality: cardinality
            subjectCardinality:
              cardinality: cardinality
            hasSubjectEntity: ""
            relationName:
            - relationName
            - relationName
            hasObjectEntity: ""
            inverseRelation: inverseRelation
            hasAttribute:
            - attributeName:
              - attributeName
              - attributeName
              hasAttributeModifier:
              - attributeModifier: attributeModifier
              - attributeModifier: attributeModifier
            - attributeName:
              - attributeName
              - attributeName
              hasAttributeModifier:
              - attributeModifier: attributeModifier
              - attributeModifier: attributeModifier
            hasRelationModifier:
            - relationModifier: relationModifier
            - relationModifier: relationModifier
          - objectCardinality:
              cardinality: cardinality
            subjectCardinality:
              cardinality: cardinality
            hasSubjectEntity: ""
            relationName:
            - relationName
            - relationName
            hasObjectEntity: ""
            inverseRelation: inverseRelation
            hasAttribute:
            - attributeName:
              - attributeName
              - attributeName
              hasAttributeModifier:
              - attributeModifier: attributeModifier
              - attributeModifier: attributeModifier
            - attributeName:
              - attributeName
              - attributeName
              hasAttributeModifier:
              - attributeModifier: attributeModifier
              - attributeModifier: attributeModifier
            hasRelationModifier:
            - relationModifier: relationModifier
            - relationModifier: relationModifier
          hasEntityModifier:
          - entityModifier: entityModifier
          - entityModifier: entityModifier
          entityName:
          - entityName
          - entityName
          entityUrl: entityUrl
          hasAttribute:
          - attributeName:
            - attributeName
            - attributeName
            hasAttributeModifier:
            - attributeModifier: attributeModifier
            - attributeModifier: attributeModifier
          - attributeName:
            - attributeName
            - attributeName
            hasAttributeModifier:
            - attributeModifier: attributeModifier
            - attributeModifier: attributeModifier
          isSubjectInRelation:
          - objectCardinality:
              cardinality: cardinality
            subjectCardinality:
              cardinality: cardinality
            hasSubjectEntity: ""
            relationName:
            - relationName
            - relationName
            hasObjectEntity: ""
            inverseRelation: inverseRelation
            hasAttribute:
            - attributeName:
              - attributeName
              - attributeName
              hasAttributeModifier:
              - attributeModifier: attributeModifier
              - attributeModifier: attributeModifier
            - attributeName:
              - attributeName
              - attributeName
              hasAttributeModifier:
              - attributeModifier: attributeModifier
              - attributeModifier: attributeModifier
            hasRelationModifier:
            - relationModifier: relationModifier
            - relationModifier: relationModifier
          - objectCardinality:
              cardinality: cardinality
            subjectCardinality:
              cardinality: cardinality
            hasSubjectEntity: ""
            relationName:
            - relationName
            - relationName
            hasObjectEntity: ""
            inverseRelation: inverseRelation
            hasAttribute:
            - attributeName:
              - attributeName
              - attributeName
              hasAttributeModifier:
              - attributeModifier: attributeModifier
              - attributeModifier: attributeModifier
            - attributeName:
              - attributeName
              - attributeName
              hasAttributeModifier:
              - attributeModifier: attributeModifier
              - attributeModifier: attributeModifier
            hasRelationModifier:
            - relationModifier: relationModifier
            - relationModifier: relationModifier
        title: title
      properties:
        title:
          title: title
          type: string
        entities:
          items:
            $ref: '#/components/schemas/Entity'
          title: entities
          type: array
      title: Partition
    Schema:
      example:
        schemaFile: schemaFile
        entities:
        - null
        - null
        implementation: implementation
        description: description
        id: id
        sourceFile: sourceFile
      properties:
        id:
          title: id
          type: string
        description:
          title: description
          type: string
        implementation:
          title: implementation
          type: string
        schemaFile:
          title: schemaFile
          type: string
        sourceFile:
          title: sourceFile
          type: string
        entities:
          items:
            $ref: '#/components/schemas/EntityLink'
          title: entities
          type: array
          xml:
            name: entities
      title: Schema
      xml:
        name: schema
    AttributeModifier:
      enum:
      - AttributeModifier: key
      example:
        attributeModifier: attributeModifier
      properties:
        attributeModifier:
          type: string
      required:
      - attributeModifier
      title: AttributeModifier
    Attribute:
      example:
        attributeName:
        - attributeName
        - attributeName
        hasAttributeModifier:
        - attributeModifier: attributeModifier
        - attributeModifier: attributeModifier
      properties:
        hasAttributeModifier:
          default: []
          items:
            $ref: '#/components/schemas/AttributeModifier'
          title: hasAttributeModifier
          type: array
        attributeName:
          items:
            type: string
          minItems: 1
          title: attributeName
          type: array
      required:
      - attributeName
      title: Attribute
    EntityModifier:
      enum:
      - EntityModifier: weak
      example:
        entityModifier: entityModifier
      properties:
        entityModifier:
          type: string
      required:
      - entityModifier
      title: EntityModifier
...
    Entity:
      example:
        isObjectInRelation:
        - objectCardinality:
            cardinality: cardinality
          subjectCardinality:
            cardinality: cardinality
          hasSubjectEntity: ""
          relationName:
          - relationName
          - relationName
          hasObjectEntity: ""
          inverseRelation: inverseRelation
          hasAttribute:
          - attributeName:
            - attributeName
            - attributeName
            hasAttributeModifier:
            - attributeModifier: attributeModifier
            - attributeModifier: attributeModifier
          - attributeName:
            - attributeName
            - attributeName
            hasAttributeModifier:
            - attributeModifier: attributeModifier
            - attributeModifier: attributeModifier
          hasRelationModifier:
          - relationModifier: relationModifier
          - relationModifier: relationModifier
        - objectCardinality:
            cardinality: cardinality
          subjectCardinality:
            cardinality: cardinality
          hasSubjectEntity: ""
          relationName:
          - relationName
          - relationName
          hasObjectEntity: ""
          inverseRelation: inverseRelation
          hasAttribute:
          - attributeName:
            - attributeName
            - attributeName
            hasAttributeModifier:
            - attributeModifier: attributeModifier
            - attributeModifier: attributeModifier
          - attributeName:
            - attributeName
            - attributeName
            hasAttributeModifier:
            - attributeModifier: attributeModifier
            - attributeModifier: attributeModifier
          hasRelationModifier:
          - relationModifier: relationModifier
          - relationModifier: relationModifier
        hasEntityModifier:
        - entityModifier: entityModifier
        - entityModifier: entityModifier
        entityName:
        - entityName
        - entityName
        entityUrl: entityUrl
        hasAttribute:
        - attributeName:
          - attributeName
          - attributeName
          hasAttributeModifier:
          - attributeModifier: attributeModifier
          - attributeModifier: attributeModifier
        - attributeName:
          - attributeName
          - attributeName
          hasAttributeModifier:
          - attributeModifier: attributeModifier
          - attributeModifier: attributeModifier
        isSubjectInRelation:
        - objectCardinality:
            cardinality: cardinality
          subjectCardinality:
            cardinality: cardinality
          hasSubjectEntity: ""
          relationName:
          - relationName
          - relationName
          hasObjectEntity: ""
          inverseRelation: inverseRelation
          hasAttribute:
          - attributeName:
            - attributeName
            - attributeName
            hasAttributeModifier:
            - attributeModifier: attributeModifier
            - attributeModifier: attributeModifier
          - attributeName:
            - attributeName
            - attributeName
            hasAttributeModifier:
            - attributeModifier: attributeModifier
            - attributeModifier: attributeModifier
          hasRelationModifier:
          - relationModifier: relationModifier
          - relationModifier: relationModifier
        - objectCardinality:
            cardinality: cardinality
          subjectCardinality:
            cardinality: cardinality
          hasSubjectEntity: ""
          relationName:
          - relationName
          - relationName
          hasObjectEntity: ""
          inverseRelation: inverseRelation
          hasAttribute:
          - attributeName:
            - attributeName
            - attributeName
            hasAttributeModifier:
            - attributeModifier: attributeModifier
            - attributeModifier: attributeModifier
          - attributeName:
            - attributeName
            - attributeName
            hasAttributeModifier:
            - attributeModifier: attributeModifier
            - attributeModifier: attributeModifier
          hasRelationModifier:
          - relationModifier: relationModifier
          - relationModifier: relationModifier
      properties:
        hasAttribute:
          default: []
          items:
            $ref: '#/components/schemas/Attribute'
          title: hasAttribute
          type: array
        hasEntityModifier:
          default: []
          items:
            $ref: '#/components/schemas/EntityModifier'
          title: hasEntityModifier
          type: array
        isObjectInRelation:
          default: []
          items:
            $ref: '#/components/schemas/Relation'
          title: isObjectInRelation
          type: array
        isSubjectInRelation:
          default: []
          items:
            $ref: '#/components/schemas/Relation'
          title: isSubjectInRelation
          type: array
        entityName:
          items:
            type: string
          minItems: 1
          title: entityName
          type: array
        entityUrl:
          nullable: true
          title: entityUrl
          type: string
      required:
      - entityName
      title: Entity
...